refactor(validate): 유효성 검사 로직 수정 및 분리
