+++ src/constants/Message.ts
@@ -8,4 +8,9 @@ export const ErrorMessage = {
   },
 };
 
+export const InfoMessage = {
+  DESCRIPTION: '메뉴 등 추가 정보를 입력해 주세요.',
+  LINK: '매장 정보를 확인할 수 있는 링크를 입력해 주세요.',
+};
+
 export const ErrorId = (id: string) => `${id}-error`;
